feat(db-arch): ctx functions and use of global db (#2378)
The first commit in a series of commits related to database architecture changes. Changes here won't have effect on builds since they are guarded with a new-db-arch feature flag.

This adds some utility methods in the context to get the global & wallet & address databases and data directories.

Also adds some concrete usages of the global db.