fix(event-streaming): initial addresses registration in utxo balance streaming (#2431)
We need to initially register the addresses we wanna get balance notification for. We used to do that in the past at coin enablement time (since event streaming config was static and doesn't change at runtime), that was removed since it doesn't make sense with runtime enabled/disabled streaming. But was never replaced with the appropriate logic to register these addresses once the streamer was enabled.

This commit registers our addresses with the event streamer on initialization so it could track them.